files/requirements.txt
pandas==2.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
streamlit==1.37.1
python-dotenv
streamlit-autorefresh